+++ js/main.js
@@ -1,1 +1,3 @@
 import './form.js';
+import './map.js';
+import './validate.js';
